requirements.txt
google-cloud-firestore
yfinance
pandas